backend/requirements.txt
fastapi==0.101.0
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 
python-jose[cryptography]
passlib
bcrypt
email-validator>=2.0.0
pydantic[email]>=2.0.0